Remove machine-code mode, it's not actually supported on BASIC cartridges
The document explaining the tape format may have actually been for the
FS-7000. Only four pages were included, so it's not clear.

Through testing, the machine-code mode does not work on either BASIC IIIa
or BASIC IIIb. Instead, machine code can be loaded using the key-code for
BASIC. The program is loaded to 0x9800.

Once loaded using the BASIC tape format, the machine-code can be run
with `CALL &H9800`